Config file integration w.i.p.
