+++ issuer_controller/config/schemas.yml
@@ -1,11 +1,55 @@
 # Documentation: https://github.com/bcgov/von-agent-template/tree/master/von-x-agent/config
 
-- name: ian-permit.ian-ville
+- name: ian-registration.ian-ville
+  version: '1.0.0'
+  description: The ian-permit credential issued by ian-cille
+  path: /ian-ville/ian-registration
+  attributes:
+    corp_num:
+      label_en: Registration ID
+      description_en: Registration/Incorporation Number or Identifying Number
+      required: true
+    registration_date:
+      label_en: Registration Date
+      description_en: Date of Registration, Incorporation, Continuation or Amalgamation
+      required: false
+    entity_name:
+      label_en: Name
+      description_en: The legal name of entity
+      required: true
+    entity_name_effective:
+      label_en: Name Effective Date
+      description_en: Date current name became effective
+      required: true
+    entity_status:
+      label_en: Registration Status
+      description_en: Status of the entity (active or historical)
+      required: true
+    entity_status_effective:
+      label_en: Status Effective Date
+      description_en: Date status became effective
+      required: true
+    entity_type:
+      label_en: Registration Type
+      description_en: Type of entity incorporated or registered
+      required: true
+    registered_jurisdiction:
+      label_en: Registered Jurisdiction
+      description_en: The jurisdiction an entity was created in
+      required: false
+    effective_date:
+      label_en: Credential Effective Date
+      description_en: Date Credential is effective
+      required: true
+    expiry_date:
+      label_en: Credential Expiry Date
+      description_en: Date Credential expired
+      required: false
 
+- name: ian-permit.ian-ville
   version: '1.0.0'
   description: The ian-permit credential issued by ian-cille
   path: /ian-ville/ian-permit
-
   attributes:
     - corp_num
     - legal_name
