+++ issuer_controller/config/services.yml
@@ -14,22 +14,6 @@ issuers:
       api_url: $TOB_API_URL
       sign_target: false
 
-    wallet:
-      name: ${POSTGRESQL_WALLET_NAME:-myorg_issuer}
-      seed: $WALLET_SEED_VONX
-      type: $WALLET_TYPE
-      params:
-        storage_config:
-          url: "$POSTGRESQL_WALLET_HOST:$POSTGRESQL_WALLET_PORT"
-      access_creds:
-        key: $WALLET_ENCRYPTION_KEY
-        storage_credentials:
-          account: $POSTGRESQL_WALLET_USER
-          password: $POSTGRESQL_WALLET_PASSWORD
-          admin_account: ${POSTGRESQL_WALLET_ADMIN_USER:-postgres}
-          admin_password: $POSTGRESQL_WALLET_ADMIN_PASSWORD
-        key_derivation_method: ARGON2I_MOD
-
     credential_types:
     - description: Permit
       schema: ian-permit.ian-ville
@@ -98,11 +82,11 @@ issuers:
 verifiers:
   bctob:
     name: BC OrgBook
+    alias: tob-agent
     connection:
       type: TheOrgBook
       api_url: $TOB_API_URL
-    wallet:
-      seed: "tob-verifier-wallet-000000000001"
+      agent_admin_url: $TOB_AGENT_ADMIN_URL
 
 proof_requests:
   greenlight_registration:
